+++ ui/app/components/send/send-content/send-amount-row/amount-max-button/amount-max-button.container.js
@@ -1,9 +1,11 @@
 import { connect } from 'react-redux'
 import {
-  getGasTotal,
   getSelectedToken,
   getSendFromBalance,
   getTokenBalance,
+  getUtxos,
+  getSelectedAddress,
+  getSelectedSlpAddress,
 } from '../../../send.selectors.js'
 import { getMaxModeOn } from './amount-max-button.selectors.js'
 import { calcMaxAmount } from './amount-max-button.utils.js'
@@ -19,10 +21,12 @@ export default connect(
 function mapStateToProps (state) {
   return {
     balance: getSendFromBalance(state),
-    gasTotal: getGasTotal(state),
     maxModeOn: getMaxModeOn(state),
     selectedToken: getSelectedToken(state),
+    selectedAddress: getSelectedAddress(state),
+    selectedSlpAddress: getSelectedSlpAddress(state),
     tokenBalance: getTokenBalance(state),
+    utxo: getUtxos(state),
   }
 }